% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_labels}
\alias{assert_labels}
\title{Assert Labels}
\usage{
assert_labels(labels)
}
\arguments{
\item{labels}{A list of labels to be used in the query.}
}
\value{
A list of labels or NULL.
}
\description{
Check that labels are constructed correctly.
}
\seealso{
Other assert: 
\code{\link{assert_endpoint}()},
\code{\link{assert_token}()}
}
\concept{assert}
