% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{toGtsInputFormat}
\alias{toGtsInputFormat}
\title{Convert a data frame}
\usage{
toGtsInputFormat(dataFrame)
}
\arguments{
\item{dataFrame}{data frame}
}
\value{
character vector
}
\description{
Convert a data frame into GTS input format. First column of the data frame must be named "timestamp". Column names corresponding to GTS values must be in the <selector> form: "classname{label1=value1,label2=value2,..}" (or "classname"). Column names corresponding to geographic information must be be in the form "<selector>.lat", "<selector>.lon" or "<selector>.elev"
}
