% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{postWarpscript}
\alias{postWarpscript}
\title{Post Warpscript Code}
\usage{
postWarpscript(warpscript, outputType = "json",
  endpoint = "http://localhost:8080/api/v0/exec", withLabels = FALSE)
}
\arguments{
\item{warpscript}{code or file name ending with .mc2}

\item{outputType}{one of:
\itemize{
  \item "json" - returns a json (default)
  \item "raw" - returns a character vector
  \item "pretty" - returns a json that is pretty indented
  \item "list" - returns a list
  \item "data.table" - returns a single data.table with all data. Requires the first level of the stack to be a list of GTS
  \item "gts.list" - returns a named list of one data.table per GTS. Requires the first level of the stack to be a list of GTS. The keys of the returned list are the GTS classnames. NaN values are not supported
  \item "dataFrame" - deprecated in favors of data.table
}}

\item{endpoint}{egress endpoint. Default to "http://localhost:8080/api/v0/exec"}

\item{withLabels}{if TRUE, labels are appended to classnames when outputType is "data.table" or "gts.list"}
}
\value{
character vector or json or named list or data.table
}
\description{
Post warpscript code to a Warp 10 instance and retrieve response as a specified type
}
