% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{toGtsInputFormat}
\alias{toGtsInputFormat}
\title{Convert a data frame}
\usage{
toGtsInputFormat(dataFrame)
}
\arguments{
\item{dataFrame}{first column must be named "timestamp". Column names corresponding to GTS values must be in the <selector> form: "classname\{label1=value1,label2=value2,...\}" (or "classname"). Optional geo columns must be named "<selector>.lat", "<selector>.lon" or "<selector>.elev"}
}
\value{
data points in GTS input format as a character vector
}
\description{
Convert a data frame into GTS input format.
}
